dag: remove `UnixfsNode` in Layout of trickledag
This patch is the part of trickledag work which is similar to the
merkledag work in commit 474b77a2bdb1c ("importer: remove `UnixfsNode`
from the balanced builder"). Two helper functions(fillTrickleRecFSNode
and FillFSNodeLayer) is introduced temporarily for modifing the Layout
functions. These two funtions will be removed when all the code of
UnixfsNode is removed in trickledag.go.

Test ipfs add and get commands to check whether get the same hash of
file after the code changes.

License: MIT
Signed-off-by: Bamvor Zhang <jian.zhang@ipfsbit.com>