Fix #27: Remove batching from importers
A batching DAG service is forced onto the
users of the importers, but they could just wrap
the given DAGSerice in the batching one to get the same
functionality (remembering to Close it at the end of the
proccess).

As detailed in #27, the importers should not be making choices
about what DAGService is the right one to use and wrapping the
given one.

This change requires wrapping the DAGService in go-ipfs into
ipld.Batch. and closing it when Finishing the adding process.